+++ updates/20200206_0006_remove_unique_constraint_from_items.php
@@ -6,9 +6,9 @@ namespace Vdlp\RssFetcher\Updates;
 
 use Illuminate\Database\Schema\Blueprint;
 use October\Rain\Database\Updates\Migration;
-use October\Rain\Support\Facades\Schema;
+use Schema;
 
-class RemoveUniqueConstraintFromItems extends Migration
+return new class extends Migration
 {
     public function up(): void
     {
@@ -23,4 +23,4 @@ class RemoveUniqueConstraintFromItems extends Migration
             $table->unique('item_id', 'item_id_unique');
         });
     }
-}
+};